% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/jiebaRD-package.r
\docType{package}
\name{jiebaRD}
\alias{jiebaRD}
\alias{jiebaRD-package}
\title{A package for Chinese text segmentation}
\description{
jiebaR is a package for Chinese text segmentation, keyword extraction
and speech tagging. This package provides the data files required by jiebaR.
jiebaR supports four types of segmentation mode: Maximum Probability, Hidden Markov Model,
Query Segment and Mix Segment.
}
\details{
You can use custom dictionary to be included in the jiebaR default dictionary.
jiebaR can also identify new words, but adding your own new words will ensure a higher
accuracy.
}
\examples{
library("jiebaRD")
}
\author{
Qin Wenfeng <\url{http://qinwenfeng.com}>
}
\references{
CppJieba \url{https://github.com/aszxqw/cppjieba};
}
\seealso{
JiebaR \url{https://github.com/qinwf/jiebaR};
}

